Consolidate MetricHandler and OperationHandler into a common parameterized implementation since they effectively do the same thing.
