+++ clustering/common/src/main/java/org/jboss/as/clustering/controller/Executable.java
@@ -28,6 +28,7 @@ import org.jboss.dmr.ModelNode;
 /**
  * Generic interface for some object capable of execution.
  * @author Paul Ferraro
+ * @param C the execution context
  */
 public interface Executable<C> {
     /**
