+++ clustering/common/src/main/java/org/jboss/as/clustering/controller/Executor.java
@@ -22,10 +22,23 @@
 
 package org.jboss.as.clustering.controller;
 
+import org.jboss.as.controller.OperationContext;
+import org.jboss.as.controller.OperationFailedException;
+import org.jboss.dmr.ModelNode;
+
 /**
- * Encapsulates the execution of a runtime operation.
+ * Encapsulates the execution of a contextual executable.
  * @author Paul Ferraro
- * @param <C> the operation execution context.
+ * @param C the execution context
+ * @param E the contextual executable
  */
-public interface OperationExecutor<C> extends Executor<C, Operation<C>> {
+public interface Executor<C, E extends Executable<C>> {
+    /**
+     * Executes the specified executable against the specified operation context.
+     * @param context an operation context
+     * @param executable the contextual executable object
+     * @return the result of the execution (possibly null).
+     * @throws OperationFailedException if execution fails
+     */
+    ModelNode execute(OperationContext context, E executable) throws OperationFailedException;
 }
